requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
opencv-python>=4.8.0
mediapipe>=0.10.0

# 3D Pose and Mesh
smplx>=0.1.28
trimesh>=3.23.0
pyrender>=0.1.45

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0
seaborn>=0.12.0

# Video Processing
moviepy>=1.0.3
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# Filtering and Signal Processing
filterpy>=1.4.5

# Configuration
pyyaml>=6.0
hydra-core>=1.3.0

# Utilities
tqdm>=4.65.0
pillow>=10.0.0
